Widen ILibGetTimeStamp() to long long
- ILibGetTimeStamp() returned long, and tv_sec * 1000 overflows a 32-bit long, which is every Windows build (x64 uses 32-bit longs) and every 32-bit target.
  That is signed overflow, which is undefined behaviour.
  Fix: return long long

+++ microstack/ILibParsers.h
@@ -231,7 +231,7 @@ char *ILibWideToUTF8_stupidEx(WCHAR* wstr, int wstrBYTESIZE, char *buffer, int b
 
 
 int ILibGetLocalTime(char *dest, int destLen);
-long ILibGetTimeStamp();
+long long ILibGetTimeStamp();
 
 #ifdef ILibEXPORT
 #ifdef WIN32
